Consolidate HostProvisioner mocks
