+++ node-repository/src/test/java/com/yahoo/vespa/hosted/provision/maintenance/DynamicProvisioningMaintainerTest.java
@@ -27,11 +27,9 @@ import com.yahoo.vespa.hosted.provision.node.Agent;
 import com.yahoo.vespa.hosted.provision.node.Allocation;
 import com.yahoo.vespa.hosted.provision.node.Generation;
 import com.yahoo.vespa.hosted.provision.node.IP;
-import com.yahoo.vespa.hosted.provision.provisioning.FatalProvisioningException;
 import com.yahoo.vespa.hosted.provision.provisioning.FlavorConfigBuilder;
-import com.yahoo.vespa.hosted.provision.provisioning.HostProvisioner;
-import com.yahoo.vespa.hosted.provision.provisioning.ProvisionedHost;
 import com.yahoo.vespa.hosted.provision.provisioning.ProvisioningTester;
+import com.yahoo.vespa.hosted.provision.testutils.MockHostProvisioner;
 import com.yahoo.vespa.hosted.provision.testutils.MockNameResolver;
 import com.yahoo.vespa.service.duper.ConfigServerApplication;
 import com.yahoo.vespa.service.duper.ConfigServerHostApplication;
@@ -39,20 +37,15 @@ import org.junit.Test;
 
 import java.time.Duration;
 import java.time.Instant;
-import java.util.ArrayList;
-import java.util.EnumSet;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
-import static com.yahoo.vespa.hosted.provision.maintenance.DynamicProvisioningMaintainerTest.MockHostProvisioner.Behaviour;
+import static com.yahoo.vespa.hosted.provision.testutils.MockHostProvisioner.Behaviour;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertSame;
 import static org.junit.Assert.assertTrue;
 
 /**
@@ -115,7 +108,7 @@ public class DynamicProvisioningMaintainerTest {
 
         tester.maintainer.maintain();
         assertTrue("Failed host is deprovisioned", tester.nodeRepository.nodes().node(failedHost.get().hostname()).isEmpty());
-        assertEquals(1, tester.hostProvisioner.deprovisionedHosts);
+        assertEquals(1, tester.hostProvisioner.deprovisionedHosts());
     }
 
     @Test
@@ -126,7 +119,7 @@ public class DynamicProvisioningMaintainerTest {
                                                new ClusterCapacity(1, 16, 24, 100, 1.0)),
                                        ClusterCapacity.class);
 
-        assertEquals(0, tester.hostProvisioner.provisionedHosts.size());
+        assertEquals(0, tester.hostProvisioner.provisionedHosts().size());
         assertEquals(11, tester.nodeRepository.nodes().list().size());
         assertTrue(tester.nodeRepository.nodes().node("host2").isPresent());
         assertTrue(tester.nodeRepository.nodes().node("host2-1").isPresent());
@@ -136,7 +129,7 @@ public class DynamicProvisioningMaintainerTest {
 
         tester.maintainer.maintain();
 
-        assertEquals(2, tester.hostProvisioner.provisionedHosts.size());
+        assertEquals(2, tester.hostProvisioner.provisionedHosts().size());
         assertEquals(2, tester.provisionedHostsMatching(new NodeResources(48, 128, 1000, 10)));
         NodeList nodesAfter = tester.nodeRepository.nodes().list();
         assertEquals(11, nodesAfter.size());  // 2 removed, 2 added
@@ -151,13 +144,13 @@ public class DynamicProvisioningMaintainerTest {
     public void preprovision_with_shared_host() {
         var tester = new DynamicProvisioningTester().addInitialNodes();
         // Makes provisioned hosts 48-128-1000-10
-        tester.hostProvisioner.provisionSharedHost("host4");
+        tester.hostProvisioner.overrideHostFlavor("host4");
 
         tester.flagSource.withListFlag(PermanentFlags.PREPROVISION_CAPACITY.id(),
                 List.of(new ClusterCapacity(2, 1, 30, 20, 3.0)),
                 ClusterCapacity.class);
 
-        assertEquals(0, tester.hostProvisioner.provisionedHosts.size());
+        assertEquals(0, tester.hostProvisioner.provisionedHosts().size());
         assertEquals(11, tester.nodeRepository.nodes().list().size());
         assertTrue(tester.nodeRepository.nodes().node("host2").isPresent());
         assertTrue(tester.nodeRepository.nodes().node("host2-1").isPresent());
@@ -196,7 +189,7 @@ public class DynamicProvisioningMaintainerTest {
 
         tester.maintainer.maintain();
 
-        assertEquals(2, tester.hostProvisioner.provisionedHosts.size());
+        assertEquals(2, tester.hostProvisioner.provisionedHosts().size());
         assertEquals(2, tester.provisionedHostsMatching(new NodeResources(48, 128, 1000, 10)));
         assertEquals(10, tester.nodeRepository.nodes().list().size());  // 3 removed, 2 added
         assertTrue("preprovision capacity is prefered on shared hosts", tester.nodeRepository.nodes().node("host3").isEmpty());
@@ -212,7 +205,7 @@ public class DynamicProvisioningMaintainerTest {
         tester.maintainer.maintain();
 
         assertEquals("one provisioned host has been deprovisioned, so there are 2 -> 1 provisioned hosts",
-                1, tester.hostProvisioner.provisionedHosts.size());
+                     1, tester.hostProvisioner.provisionedHosts().size());
         assertEquals(1, tester.provisionedHostsMatching(new NodeResources(48, 128, 1000, 10)));
         assertEquals(9, tester.nodeRepository.nodes().list().size());  // 4 removed, 2 added
         if (tester.nodeRepository.nodes().node("hostname100").isPresent()) {
@@ -226,7 +219,7 @@ public class DynamicProvisioningMaintainerTest {
     }
 
     private void verifyFirstMaintain(DynamicProvisioningTester tester) {
-        assertEquals(1, tester.hostProvisioner.provisionedHosts.size());
+        assertEquals(1, tester.hostProvisioner.provisionedHosts().size());
         assertEquals(1, tester.provisionedHostsMatching(new NodeResources(48, 128, 1000, 10)));
         assertEquals(10, tester.nodeRepository.nodes().list().size());  // 2 removed, 1 added
         assertTrue("Failed host 'host2' is deprovisioned", tester.nodeRepository.nodes().node("host2").isEmpty());
@@ -266,17 +259,17 @@ public class DynamicProvisioningMaintainerTest {
 
     private void assertWithMinCount(int minCount, int provisionCount, int deprovisionCount) {
         var tester = new DynamicProvisioningTester().addInitialNodes();
-        tester.hostProvisioner.provisionSharedHost("host4");
+        tester.hostProvisioner.overrideHostFlavor("host4");
 
         tester.flagSource.withJacksonFlag(PermanentFlags.SHARED_HOST.id(), new SharedHost(null, minCount), SharedHost.class);
         tester.maintainer.maintain();
-        assertEquals(provisionCount, tester.hostProvisioner.provisionedHosts.size());
-        assertEquals(deprovisionCount, tester.hostProvisioner.deprovisionedHosts);
+        assertEquals(provisionCount, tester.hostProvisioner.provisionedHosts().size());
+        assertEquals(deprovisionCount, tester.hostProvisioner.deprovisionedHosts());
 
         // Verify next maintain is a no-op
         tester.maintainer.maintain();
-        assertEquals(provisionCount, tester.hostProvisioner.provisionedHosts.size());
-        assertEquals(deprovisionCount, tester.hostProvisioner.deprovisionedHosts);
+        assertEquals(provisionCount, tester.hostProvisioner.provisionedHosts().size());
+        assertEquals(deprovisionCount, tester.hostProvisioner.deprovisionedHosts());
     }
 
     @Test
@@ -300,14 +293,14 @@ public class DynamicProvisioningMaintainerTest {
 
         // Hosts are provisioned
         assertEquals(2, tester.provisionedHostsMatching(resources1));
-        assertEquals(0, tester.hostProvisioner.deprovisionedHosts);
+        assertEquals(0, tester.hostProvisioner.deprovisionedHosts());
 
         // Next maintenance run does nothing
         tester.assertNodesUnchanged();
 
         // Pretend shared-host flag has been set to host4's flavor
         var sharedHostNodeResources = new NodeResources(48, 128, 1000, 10, NodeResources.DiskSpeed.fast, NodeResources.StorageType.remote);
-        tester.hostProvisioner.provisionSharedHost("host4");
+        tester.hostProvisioner.overrideHostFlavor("host4");
 
         // Next maintenance run does nothing
         tester.assertNodesUnchanged();
@@ -456,7 +449,7 @@ public class DynamicProvisioningMaintainerTest {
 
         public DynamicProvisioningTester(Cloud cloud) {
             MockNameResolver nameResolver = new MockNameResolver();
-            this.hostProvisioner = new MockHostProvisioner(flavors, nameResolver);
+            this.hostProvisioner = new MockHostProvisioner(flavors.getFlavors(), nameResolver, 0);
             this.provisioningTester = new ProvisioningTester.Builder().zone(new Zone(cloud, SystemName.defaultSystem(),
                                                                                      Environment.defaultEnvironment(),
                                                                                      RegionName.defaultName()))
@@ -529,124 +522,15 @@ public class DynamicProvisioningMaintainerTest {
         }
 
         private long provisionedHostsMatching(NodeResources resources) {
-            return hostProvisioner.provisionedHosts.stream()
-                                                   .filter(host -> host.generateHost().resources().compatibleWith(resources))
-                                                   .count();
+            return hostProvisioner.provisionedHosts().stream()
+                                  .filter(host -> host.generateHost().resources().compatibleWith(resources))
+                                  .count();
         }
 
         private void assertNodesUnchanged() {
             NodeList nodes = nodeRepository.nodes().list();
             maintainer.maintain();
             assertEquals("Nodes are unchanged after maintenance run", nodes, nodeRepository.nodes().list());
-        }
-
-    }
-
-    static class MockHostProvisioner implements HostProvisioner {
-
-        private final List<ProvisionedHost> provisionedHosts = new ArrayList<>();
-        private final NodeFlavors flavors;
-        private final MockNameResolver nameResolver;
-
-        private int deprovisionedHosts = 0;
-        private EnumSet<Behaviour> behaviours = EnumSet.noneOf(Behaviour.class);
-        private Optional<Flavor> provisionHostFlavor = Optional.empty();
-
-        public MockHostProvisioner(NodeFlavors flavors, MockNameResolver nameResolver) {
-            this.flavors = flavors;
-            this.nameResolver = nameResolver;
-        }
-
-        public MockHostProvisioner provisionSharedHost(String flavorName) {
-            provisionHostFlavor = Optional.of(flavors.getFlavorOrThrow(flavorName));
-            return this;
-        }
-
-        @Override
-        public List<ProvisionedHost> provisionHosts(List<Integer> provisionIndexes, NodeResources resources,
-                                                    ApplicationId applicationId, Version osVersion, HostSharing sharing) {
-            Flavor hostFlavor = provisionHostFlavor
-                    .orElseGet(() -> flavors.getFlavors().stream()
-                            .filter(f -> !f.isDocker())
-                            .filter(f -> f.resources().compatibleWith(resources))
-                            .findFirst()
-                            .orElseThrow(() -> new IllegalArgumentException("No host flavor found satisfying " + resources)));
-
-            List<ProvisionedHost> hosts = new ArrayList<>();
-            for (int index : provisionIndexes) {
-                hosts.add(new ProvisionedHost("host" + index,
-                                              "hostname" + index,
-                                              hostFlavor,
-                                              Optional.empty(),
-                                              createAddressesForHost(hostFlavor, index),
-                                              resources,
-                                              osVersion));
-            }
-            provisionedHosts.addAll(hosts);
-            return hosts;
-        }
-
-        private List<Address> createAddressesForHost(Flavor flavor, int hostIndex) {
-            long numAddresses = Math.max(1, Math.round(flavor.resources().bandwidthGbps()));
-            return IntStream.range(0, (int) numAddresses)
-                    .mapToObj(i -> new Address("nodename" + hostIndex + "_" + i))
-                    .collect(Collectors.toList());
-        }
-
-        @Override
-        public List<Node> provision(Node host, Set<Node> children) throws FatalProvisioningException {
-            if (behaviours.contains(Behaviour.failProvisioning)) throw new FatalProvisioningException("Failed to provision node(s)");
-            assertSame(Node.State.provisioned, host.state());
-            List<Node> result = new ArrayList<>();
-            result.add(withIpAssigned(host));
-            for (var child : children) {
-                assertSame(Node.State.reserved, child.state());
-                result.add(withIpAssigned(child));
-            }
-            return result;
-        }
-
-        @Override
-        public void deprovision(Node host) {
-            if (behaviours.contains(Behaviour.failDeprovisioning)) throw new FatalProvisioningException("Failed to deprovision node");
-            provisionedHosts.removeIf(provisionedHost -> provisionedHost.hostHostname().equals(host.hostname()));
-            deprovisionedHosts++;
-        }
-
-        private MockHostProvisioner with(Behaviour first, Behaviour... rest) {
-            this.behaviours = EnumSet.of(first, rest);
-            return this;
-        }
-
-        private MockHostProvisioner without(Behaviour first, Behaviour... rest) {
-            Set<Behaviour> behaviours = new HashSet<>(this.behaviours);
-            behaviours.removeAll(EnumSet.of(first, rest));
-            this.behaviours = behaviours.isEmpty() ? EnumSet.noneOf(Behaviour.class) : EnumSet.copyOf(behaviours);
-            return this;
-        }
-
-        private Node withIpAssigned(Node node) {
-            if (node.parentHostname().isPresent()) return node;
-            int hostIndex = Integer.parseInt(node.hostname().replaceAll("^[a-z]+|-\\d+$", ""));
-            Set<String> addresses = Set.of("::" + hostIndex + ":0");
-            Set<String> ipAddressPool = new HashSet<>();
-            if (!behaviours.contains(Behaviour.failDnsUpdate)) {
-                nameResolver.addRecord(node.hostname(), addresses.iterator().next());
-                for (int i = 1; i <= 2; i++) {
-                    String ip = "::" + hostIndex + ":" + i;
-                    ipAddressPool.add(ip);
-                    nameResolver.addRecord(node.hostname() + "-" + i, ip);
-                }
-            }
-
-            IP.Pool pool = node.ipConfig().pool().withIpAddresses(ipAddressPool);
-            return node.with(node.ipConfig().withPrimary(addresses).withPool(pool));
-        }
-
-        enum Behaviour {
-            failProvisioning,
-            failDeprovisioning,
-            failDnsUpdate,
         }
 
     }
